:truck: 대규모 클래스 이동
