--- src/main/java/com/kyungmin/lavanderia/auth/jwt/data/repository/RefreshRepository.java
@@ -1,13 +1,0 @@
-package com.kyungmin.lavanderia.auth.jwt.data.repository;
-
-import com.kyungmin.lavanderia.auth.jwt.data.entity.RefreshEntity;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RefreshRepository extends CrudRepository<RefreshEntity,String> {
-
-    Boolean existsByRefresh(String refresh);
-
-    RefreshEntity findByRefresh(String refresh);
-}
