--- src/main/java/com/kyungmin/lavanderia/auth/jwt/service/ReissueService.java
@@ -1,9 +1,0 @@
-package com.kyungmin.lavanderia.auth.jwt.service;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.ResponseEntity;
-
-public interface ReissueService {
-    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response);
-}
