--- src/main/java/com/kyungmin/lavanderia/auth/oauth2/data/dto/NaverResponse.java
@@ -1,35 +1,0 @@
-package com.kyungmin.lavanderia.auth.oauth2.data.dto;
-
-import java.util.Map;
-
-public class NaverResponse implements OAuth2Response{
-    private final Map<String, Object> attribute;
-
-    public NaverResponse(Map<String, Object> attribute) {
-        this.attribute = (Map<String, Object>) attribute.get("response");
-    }
-
-    @Override
-    public String getProvider() {
-
-        return "naver";
-    }
-
-    @Override
-    public String getProviderId() {
-
-        return attribute.get("id").toString();
-    }
-
-    @Override
-    public String getEmail() {
-
-        return attribute.get("email").toString();
-    }
-
-    @Override
-    public String getName() {
-
-        return attribute.get("name").toString();
-    }
-}
