--- src/main/java/com/kyungmin/lavanderia/auth/util/MakeCookie.java
@@ -1,19 +1,0 @@
-package com.kyungmin.lavanderia.auth.util;
-
-import jakarta.servlet.http.Cookie;
-import org.springframework.stereotype.Component;
-
-@Component
-public class MakeCookie {
-
-    public Cookie createCookie(String key, String value) {
-
-        Cookie cookie = new Cookie(key, value);
-        cookie.setMaxAge(24*60*60);
-        //cookie.setSecure(true);
-        //cookie.setPath("/");
-        cookie.setHttpOnly(true);
-
-        return cookie;
-    }
-}
