--- src/main/java/com/kyungmin/lavanderia/auth/util/TokenExpirationTime.java
@@ -1,6 +1,0 @@
-package com.kyungmin.lavanderia.auth.util;
-
-public class TokenExpirationTime {
-    public static final Long ACCESS_TIME = 600000L;
-    public static final Long REFRESH_TIME = 86400000L;  // 86400L
-}
