--- src/main/java/com/kyungmin/lavanderia/config/DatabaseConfig.java
@@ -1,36 +1,0 @@
-package com.kyungmin.lavanderia.config;
-
-
-import com.zaxxer.hikari.HikariDataSource;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-
-@Configuration
-@PropertySource("classpath:database.properties")
-@ComponentScan(basePackages = {"com.kyungmin.lavanderia.*.data.*","com.kyungmin.lavanderia.*"})
-public class DatabaseConfig {
-
-    // mysql 설정
-    @Value("${db.driverClass}")
-    private String driverClass;
-    @Value("${db.url}")
-    private String url;
-    @Value("${db.username}")
-    private String username;
-    @Value("${db.password}")
-    private String password;
-
-    @Bean
-    public HikariDataSource dataSource() {
-        HikariDataSource dataSource = new HikariDataSource();
-        dataSource.setDriverClassName(driverClass);
-        dataSource.setJdbcUrl(url);
-        dataSource.setUsername(username);
-        dataSource.setPassword(password);
-        return dataSource;
-    }
-
-}
