--- src/main/java/com/kyungmin/lavanderia/config/GoogleCloudStorageConfig.java
@@ -1,27 +1,0 @@
-package com.kyungmin.lavanderia.config;
-
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.cloud.storage.Storage;
-import com.google.cloud.storage.StorageOptions;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.io.ClassPathResource;
-
-import java.io.IOException;
-
-@Configuration
-public class GoogleCloudStorageConfig {
-
-    @Bean
-    public Storage storage() throws IOException {
-
-        ClassPathResource resource = new ClassPathResource("lavanderia-421805-157140079398.json");
-        GoogleCredentials credentials = GoogleCredentials.fromStream(resource.getInputStream());
-        String projectId = "lavanderia-421805";
-        return StorageOptions.newBuilder()
-                .setProjectId(projectId)
-                .setCredentials(credentials)
-                .build()
-                .getService();
-    }
-}
