+++ src/main/java/com/kyungmin/lavanderia/global/auth/jwt/controller/ReissueController.java
@@ -1,0 +1,23 @@
+package com.kyungmin.lavanderia.global.auth.jwt.controller;
+
+import com.kyungmin.lavanderia.global.auth.jwt.service.ReissueService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ReissueController {
+
+    private final ReissueService reissueService;
+
+    public ReissueController(ReissueService reissueService) {
+        this.reissueService = reissueService;
+    }
+
+    @PostMapping("/reissue")
+    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) {
+        return reissueService.reissue(request, response);
+    }
+}
